+++ CHANGELOG.md
@@ -95,6 +95,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Fixed kayak paddle and minecart wrench not being drawn when starting a level.
 * Fixed Lara sinking into the floor or being unable to pick up items underwater in low tunnels.
 * Fixed SHOOT_SWITCH2 shattering without OCB 444.
+* Fixed MULTIPLE_BOULDERS stopping a sector short of obstacles or immediately when placed above the floor.
 
 ### Lua API changes